Migrate hors_concours flight test to NokoGiri
